build.sh

+++ build.sh
@@ -5,7 +5,16 @@ mkdir -p ../build
 javac $(find . -name "*.java") -d build
 
 if [[ $1 != '0' ]]; then
-  echo -e "\n\033[32mExecute result\033[0m"
-  java -cp $(pwd)/build/ $(basename $1 .class)
-  exit 0
+  if [[ -f $1 ]]; then
+    echo -e "\n\033[32mExecute result\033[0m"
+    java -cp $(pwd)/build/ $(basename $1 .class)
+    exit 0
+  else
+    echo "public class {
+  public static void main(String args[]) {
+    
+  }
+}" >  $1
+    subl $1
+  fi
 fi
